Move relevant files to vision.
