--- vision/new/loader.h
@@ -1,6 +1,0 @@
-#ifndef SHADER_HPP
-#define SHADER_HPP
-#include <GL/gl.h>
-GLuint LoadShaders(const char * vertex_file_path, const char * fragment_file_path);
-
-#endif
