--- vision/new/vertex.glsl
@@ -1,7 +1,0 @@
-#version 330 core
-layout(location = 0) in vec3 vertexPosition_modelspace;
-
-void main(){
-    gl_Position.xyz = vertexPosition_modelspace;
-    gl_Position.w = 1.0;
-}
